Abandon previous logic, implement simpler calculator
The previous atempt, using an array for memory, was making it too
difficult to identify what to do, and when.

Now there are only four memory positions (2 for numbers, 2 for
functions), and a selector to "usher" inputs to the correct positions.

Also, the idea of inferred numbers (exemple: 2 + === returns 8) was
dropped, at least for now.